pc-side/requirements.txt
paho-mqtt
mysql-connector-python
PyYAML